refactor and cleanup code
